requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
moviepy>=1.0.3
sounddevice>=0.4.6
pywin32>=306
